# tpl: Support slices of uncomparable types in union and intersect