This PR continues to use a map for comparable types but adds a
slower option for comparing uncomparable types, such as maps and slices.
    
Fixes #3820