@bep 2017-10-09
What is the performance effect of this change?