@moorereason 2017-10-09
It's slower than a map lookup for sure, but we're talking microseconds.  The memory performance is a little better than master.

Given the benchmarks from my newly submitted PR #3954:

```
$ benchstat bench.master.txt bench.iss3820.txt
name         old time/op    new time/op    delta
Intersect-2    9.08µs ± 1%   13.52µs ± 2%   +48.81%  (p=0.000 n=10+10)
Union-2        7.29µs ± 1%   21.51µs ± 1%  +195.06%  (p=0.000 n=10+9)

name         old alloc/op   new alloc/op   delta
Intersect-2      544B ± 0%      336B ± 0%   -38.24%  (p=0.000 n=10+10)
Union-2          912B ± 0%      704B ± 0%   -22.81%  (p=0.000 n=10+10)

name         old allocs/op  new allocs/op  delta
Intersect-2      12.0 ± 0%      10.0 ± 0%   -16.67%  (p=0.000 n=10+10)
Union-2          25.0 ± 0%      23.0 ± 0%    -8.00%  (p=0.000 n=10+10)
```