@bep 2017-10-09
>It's slower than a map lookup for sure, but we're talking microseconds.

Sure, when n is small. You should try setting up some sub benchmarks with increasing sizes. I don't like this and we are not merging this as is. I

I think this union of slices/maps is a weird use case. I haven't looked too hard on this case (if it is real), but I suggest that we either:

1. Throw an "unsupported error" type of error in these cases.
2. If we in a simple way can do it: Handle this as a special case, i.e. use map