@bep 2017-10-09
Just to make myself clear: The number one use case for both of these methods are to intersect/union Pages (i.e. slices of *Page). So the benchmarks and implementation should reflect that.