@moorereason 2017-11-28
Replaced the commits in the PR with a different approach.  See what you think about it.  On my old system, it's comparable to master:

```text
$ benchstat bench-master.txt bench-iss3820.txt 
name         old time/op    new time/op    delta
Intersect-2    42.9µs ± 1%    43.1µs ± 1%  +0.62%  (p=0.035 n=10+10)
Union-2        17.3µs ± 1%    17.7µs ± 1%  +2.18%  (p=0.000 n=9+9)

name         old alloc/op   new alloc/op   delta
Intersect-2      544B ± 0%      544B ± 0%    ~     (all equal)
Union-2        1.63kB ± 0%    1.63kB ± 0%    ~     (all equal)

name         old allocs/op  new allocs/op  delta
Intersect-2      12.0 ± 0%      12.0 ± 0%    ~     (all equal)
Union-2          31.0 ± 0%      31.0 ± 0%    ~     (all equal)
```

```text
$ go test -bench=.* ./tpl/collections/ -benchmem
BenchmarkIntersect-2               30000             43717 ns/op             544 B/op         12 allocs/op
BenchmarkIntersectSlices-2         20000             97996 ns/op           11008 B/op        468 allocs/op
BenchmarkUnion-2                  100000             17322 ns/op            1626 B/op         31 allocs/op
BenchmarkUnionSlices-2             20000             65281 ns/op            6608 B/op        288 allocs/op
```