@moorereason 2018-02-24
Rebased.