@moorereason 2018-06-25
@bep,
Is this PR a keeper?