@bep 2018-08-15
@moorereason I have looked at this again. What #3820 needs is a better error message "slice not supported" or something. There may also be other "non-hashable" Go types that we should support (I have not seen any examples of that). But this PR introduces what looks like quadratic runtime that I think will be more surprising to people who accidentally tries to union large page slices.